Extract apps/server logic into terraform modules
We'll see if this will come back to bite me later if I want to reach
across module boundaries. But for now, separating the apps from the
flynn-server config will reduce the likelihood of making an inadvertent
change to other apps.

Also, I suspect it will let me template config, i.e. when testing a new
configuration for a flynn server I can just invoke the module a second
time with a different CLUSTER_DOMAIN or something like that.